requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
audioread==3.0.1
blinker==1.8.2
Brotli==1.1.0
click==8.1.7
dacite==1.8.1
Deprecated==1.2.14
fastapi==0.103.2
Flask==3.0.3
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
jaconv==0.4.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lyricsgenius==3.0.1
msgpack==1.0.8
mutagen==1.47.0
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
pooch==1.8.2
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pykakasi==2.3.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytube==15.0.0
rapidfuzz==3.9.7
redis==5.0.8
regex==2024.7.24
scikit-learn==1.5.1
scipy==1.14.1
soundcloud-v2==1.6.0
soundfile==0.12.1
soxr==0.5.0.post1
spotdl==4.2.7
spotipy==2.24.0
starlette==0.27.0
syncedlyrics==1.0.1
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.66.5
Unidecode==1.3.8
uvicorn==0.23.2
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
yt-dlp==2024.8.6
ytmusicapi==1.8.1
